First public commit. Added version 1.0 of skin "Khillasaurus - Mashup"
This is the first skin I am committing publicly. It is a skin called
"Khillasaurus - Mashup" and is a combination of multiple skins and my
work combined.
It is a work in progress (version 1.0), but is currently useable for
competitive play.
Next update will update legal terms, and will be found in the README.
For now, this skin is purely for personal use, as it contains some
images from other skinners.
The vast majority of the scripting was written by me, and in a very soon
update should contain content purely written by me.